[dev] finished new package `obs_gen`
finished splitting the module to the package structure
