+++ dev/obs/obs_codegen/__init__.py
@@ -1,0 +1,1 @@
+from . import c
