+++ dev/obs/obs_codegen/c.py
@@ -1,0 +1,5 @@
+"""
+Constants
+"""
+
+ROUND_THRESHOLD = 9
